requirements.txt
-e .
aws-cdk.core
aws-cdk.aws-codecommit
aws-cdk.aws-codebuild
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codedeploy
aws-cdk.aws-s3
aws-cdk.aws-ec2